holbertonhackaton2/ai-classifier/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
xgboost>=3.2.0
scikit-learn>=1.8.0
numpy==1.26.4
pandas==2.2.0
tensorflow-cpu==2.16.1
